EcoEvoModelZoo

Documentation for EcoEvoModelZoo.

EcoEvoModelZoo.AkessonModel

EcoEvoModelZoo.EcosystemModelMcCann

EcoEvoModelZoo.EcosystemModelOmnivory

EcoEvoModelZoo.ResourceCompetition

EcoEvoModelZoo.ResourceCompetitionSmoothMin

EcoEvoModelZoo.Temp

EcoEvoModelZoo.eqs!

EcoEvoModelZoo.funcresp!

EcoEvoModelZoo.generate_network

EcoEvoModelZoo.init_akesson_model

EcoEvoModelZoo.smoothstep

EcoEvoModelZoo.AkessonModel — Type

This model is inspired from Akesson et al. 2021. General version of the model, corresponding to original R script.

Specialized variants are provided in other AbstractModels.

EcoEvoModelZoo.EcosystemModelMcCann — Type

EcosystemModelMcKann

This model is inspired from McCann 1994. Similar to EcosystemModelOmnivory, but without monivory.

EcoEvoModelZoo.EcosystemModelOmnivory — Type

EcosystemModelOmnivory

This model is inspired from McCann 1997.

EcoEvoModelZoo.ResourceCompetition — Type

This model is inspired from Huisman et al. 1999 Nature.

EcoEvoModelZoo.ResourceCompetitionSmoothMin — Type

This model is inspired from Huisman et al. 1999 Nature., but where Leibig's law is replaced by imperfect substituable resources (smooth minimum). The smooth min function is parametrized by s, which is a trainable parameter.

EcoEvoModelZoo.Temp — Method

Temperature as a function of space (x), time (t), and some climate parameters

EcoEvoModelZoo.eqs! — Method

eqs(u, pars, t)

Right-hand side of dynamical equations Input:

time: Time at which function is evaluated (explicit time-dependence)

u: Vector of u variables, with 2SL entries, where S is the number

of species and L the number of patches. The first SL entries are the densities, the second SL entries are the trait means.

pars: Model parameters, given as members of a list

Output:

The derivatives of the densities and trait means, as a vector in a list */

EcoEvoModelZoo.funcresp! — Method

funcresp(n::Vector{<:Real}, Th::Vector{<:Real}, arate::Vector{<:Real}, W::Matrix{<:Real}) Type II functional response Input:

n: Vector of population densities of all species in a given patch

Th: Vector of handling times (with dummy values for resource species)

arate: Vector of attack rates (with dummy values for resource species)

W: Adjacency matrix of trophic network; W(i,j)=1 if i eats j and 0 otherwise

Output:

A matrix F(i,j), the feeding rate of consumer i on resource j

EcoEvoModelZoo.generate_network — Method

generate_network(SR::Int, SC::Int)

return matrix W[i,j], which is nonzero if consumer i eats resource j SR is the number of resource, SC the number of consumer species.

The bipartite network is generated as follows. First, both resources and consumers are labeled consecutively, based on their initial temperature adaptations: resource 1 / consumer 1 are the most cold-adapted, and resource S / consumer S the most warm-adapted. Next, we always put a feeding link between consumer i and resource i. Finally, each consumer is randomly linked to four other resource species. This yields a feeding network where every consumer is connected to five resources altogether.

Note: it seems that the definition of W_ij is that it determines which resource i is eat by consumer j

EcoEvoModelZoo.init_akesson_model — Method

init_akesson_model(
;
    tE,
    S,
    L,
    vbar,
    dbar,
    model_type,
    kappa,
    eta,
    nmin,
    aw,
    bw,
    Tmax,
    Tmin,
    Cmax,
    Cmin
)

Returns a Dictionary with all params for Akesson model.

Arguments

model_type: can take values "normal", "Tdep", "Tdep_trophic" and "trophic".

EcoEvoModelZoo.smoothstep — Method

Apply twice continuously differentiable smoothed step function to a number x Input: x: Distance from pole, measured in units of the pole-to-equator distance Output: 0 if x < 0; 10x^3-15x^4+6*x^5 if 0 <= x <= 1; otherwise 1